# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- jwtAuth: []
paths:
  /admin/user:
    get:
      tags:
      - admin-user-controller
      operationId: getAll
      parameters:
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 10
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedModelAdminUserView"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                oneOf:
                - $ref: "#/components/schemas/DtoError"
                - $ref: "#/components/schemas/ParamError"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
          content:
            text/plain:
              schema:
                type: string
    put:
      tags:
      - admin-user-controller
      operationId: edit
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditUserRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminUserView"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                oneOf:
                - $ref: "#/components/schemas/DtoError"
                - $ref: "#/components/schemas/ParamError"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
          content:
            text/plain:
              schema:
                type: string
    post:
      tags:
      - admin-user-controller
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminUserView"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                oneOf:
                - $ref: "#/components/schemas/DtoError"
                - $ref: "#/components/schemas/ParamError"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
          content:
            text/plain:
              schema:
                type: string
    delete:
      tags:
      - admin-user-controller
      operationId: delete
      parameters:
      - name: id
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                oneOf:
                - $ref: "#/components/schemas/DtoError"
                - $ref: "#/components/schemas/ParamError"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
          content:
            text/plain:
              schema:
                type: string
  /user/card/transfer:
    post:
      tags:
      - user-card-controller
      operationId: transfer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferRequest"
        required: true
      responses:
        "200":
          description: OK
  /user/card/block:
    post:
      tags:
      - user-card-controller
      operationId: requestBlock
      parameters:
      - name: cardId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /auth/login:
    post:
      tags:
      - auth-controller
      operationId: login
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                  minLength: 1
                password:
                  type: string
                  minLength: 1
                client_id:
                  type: string
                client_secret:
                  type: string
                grant_type:
                  type: string
              required:
              - password
              - username
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties: {}
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                oneOf:
                - $ref: "#/components/schemas/DtoError"
                - $ref: "#/components/schemas/ParamError"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
          content:
            text/plain:
              schema:
                type: string
  /admin/card:
    get:
      tags:
      - admin-card-controller
      operationId: getAll_1
      parameters:
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 10
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedModelAdminCardView"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                oneOf:
                - $ref: "#/components/schemas/DtoError"
                - $ref: "#/components/schemas/ParamError"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
          content:
            text/plain:
              schema:
                type: string
    post:
      tags:
      - admin-card-controller
      operationId: create_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCardRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminCardView"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                oneOf:
                - $ref: "#/components/schemas/DtoError"
                - $ref: "#/components/schemas/ParamError"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
          content:
            text/plain:
              schema:
                type: string
    delete:
      tags:
      - admin-card-controller
      operationId: delete_1
      parameters:
      - name: id
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                oneOf:
                - $ref: "#/components/schemas/DtoError"
                - $ref: "#/components/schemas/ParamError"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
          content:
            text/plain:
              schema:
                type: string
  /admin/card/block/block:
    post:
      tags:
      - block-card-controller
      operationId: block
      parameters:
      - name: cardId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                oneOf:
                - $ref: "#/components/schemas/DtoError"
                - $ref: "#/components/schemas/ParamError"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
          content:
            text/plain:
              schema:
                type: string
  /admin/card/activate:
    post:
      tags:
      - admin-card-controller
      operationId: activate
      parameters:
      - name: id
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                oneOf:
                - $ref: "#/components/schemas/DtoError"
                - $ref: "#/components/schemas/ParamError"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
          content:
            text/plain:
              schema:
                type: string
  /user/card:
    get:
      tags:
      - user-card-controller
      operationId: getMyCards
      parameters:
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 10
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          items:
            type: string
      - name: last4Digits
        in: query
        schema:
          type: string
      - name: ownerText
        in: query
        schema:
          type: string
      - name: expiration
        in: query
        schema:
          type: string
          format: date
      - name: status
        in: query
        schema:
          type: string
          enum:
          - ACTIVE
          - BLOCKED
          - EXPIRED
      - name: balance
        in: query
        schema:
          type: number
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedModelUserCardView"
  /user/card/balance:
    get:
      tags:
      - user-card-controller
      operationId: getBalance
      parameters:
      - name: cardId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: number
  /admin/card/block/pending:
    get:
      tags:
      - block-card-controller
      operationId: getPendingBlockRequests
      parameters:
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 10
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedModelPendingBlockCardView"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                oneOf:
                - $ref: "#/components/schemas/DtoError"
                - $ref: "#/components/schemas/ParamError"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
          content:
            text/plain:
              schema:
                type: string
components:
  schemas:
    EditUserRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
        password:
          type: string
          maxLength: 50
          minLength: 5
      required:
      - id
      - password
    DtoError:
      type: object
      properties:
        fields:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        otherErrors:
          type: array
          items:
            type: string
    ParamError:
      type: object
      properties:
        errors:
          type: object
          additionalProperties:
            type: string
    AdminUserView:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
    TransferRequest:
      type: object
      properties:
        source:
          type: integer
          format: int64
        target:
          type: integer
          format: int64
        amount:
          type: number
      required:
      - amount
      - source
      - target
    CreateUserRequest:
      type: object
      properties:
        username:
          type: string
          maxLength: 50
          minLength: 5
        password:
          type: string
          maxLength: 50
          minLength: 5
        role:
          type: string
          enum:
          - ADMIN
          - USER
      required:
      - password
      - username
    CreateCardRequest:
      type: object
      properties:
        ownerId:
          type: integer
          format: int64
        ownerText:
          type: string
          maxLength: 128
          minLength: 0
        expiration:
          type: string
          format: date
        balance:
          type: number
      required:
      - balance
      - expiration
      - ownerId
      - ownerText
    AdminCardView:
      type: object
      properties:
        id:
          type: integer
          format: int64
        maskedNumber:
          type: string
        ownerId:
          type: integer
          format: int64
        ownerText:
          type: string
        expiration:
          type: string
          format: date
        status:
          type: string
          enum:
          - ACTIVE
          - BLOCKED
          - EXPIRED
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
    PagedModelUserCardView:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/UserCardView"
        page:
          $ref: "#/components/schemas/PageMetadata"
    UserCardView:
      type: object
      properties:
        id:
          type: integer
          format: int64
        maskedNumber:
          type: string
        expiration:
          type: string
          format: date
        status:
          type: string
          enum:
          - ACTIVE
          - BLOCKED
          - EXPIRED
    PagedModelAdminUserView:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/AdminUserView"
        page:
          $ref: "#/components/schemas/PageMetadata"
    PagedModelAdminCardView:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/AdminCardView"
        page:
          $ref: "#/components/schemas/PageMetadata"
    PagedModelPendingBlockCardView:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/PendingBlockCardView"
        page:
          $ref: "#/components/schemas/PageMetadata"
    PendingBlockCardView:
      type: object
      properties:
        id:
          type: integer
          format: int64
        card:
          $ref: "#/components/schemas/AdminCardView"
  securitySchemes:
    jwtAuth:
      type: oauth2
      name: jwtAuth
      flows:
        password:
          tokenUrl: /auth/login
          scopes: {}